refactor(domain): Move implementation of A1 traits for CellRange to cell_range.rs
